requirements.txt
fastapi
uvicorn[standard]
torch
transformers
pillow
requests
opencv-python
numpy